Move application tests next to tested classes
